tests: move to single binary
