+++ ethers-contract/tests/it/main.rs
@@ -1,10 +1,14 @@
-// #![allow(clippy::extra_unused_type_parameters)]
+#![allow(clippy::extra_unused_type_parameters)]
+#![cfg(feature = "abigen")]
 
-#[cfg(feature = "abigen")]
 mod abigen;
-pub(crate) mod common;
 
-#[cfg(feature = "abigen")]
-mod contract;
+mod derive;
 
 mod contract_call;
+
+#[cfg(all(not(target_arch = "wasm32"), not(feature = "celo")))]
+mod common;
+
+#[cfg(all(not(target_arch = "wasm32"), not(feature = "celo")))]
+mod contract;
