+++ ethers-middleware/tests/it/main.rs
@@ -1,0 +1,19 @@
+#![allow(clippy::extra_unused_type_parameters)]
+#![cfg(not(target_arch = "wasm32"))]
+
+mod builder;
+
+mod gas_escalator;
+
+mod gas_oracle;
+
+mod signer;
+
+#[cfg(not(feature = "celo"))]
+mod nonce_manager;
+
+#[cfg(not(feature = "celo"))]
+mod stack;
+
+#[cfg(not(feature = "celo"))]
+mod transformer;
