+++ ethers-providers/tests/it/main.rs
@@ -1,0 +1,8 @@
+#![cfg(not(target_arch = "wasm32"))]
+
+mod provider;
+
+mod txpool;
+
+#[cfg(not(feature = "celo"))]
+mod ws_errors;
